fix: 🐛 seperate routes into a config
